[PIR] rename pir cf_ops.h file to cf_op.h
