+++ paddle/pir/dialect/control_flow/ir/cf_dialect.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 #include "paddle/pir/dialect/control_flow/ir/cf_dialect.h"
-#include "paddle/pir/dialect/control_flow/ir/cf_ops.h"
+#include "paddle/pir/dialect/control_flow/ir/cf_op.h"
 
 namespace pir {
 void ControlFlowDialect::initialize() { RegisterOps<YieldOp>(); }
