+++ test/cpp/pir/control_flow_dialect/if_op_test.cc
@@ -21,7 +21,7 @@
 #include "paddle/pir/core/builtin_op.h"
 #include "paddle/pir/core/program.h"
 #include "paddle/pir/dialect/control_flow/ir/cf_dialect.h"
-#include "paddle/pir/dialect/control_flow/ir/cf_ops.h"
+#include "paddle/pir/dialect/control_flow/ir/cf_op.h"
 
 TEST(if_op_test, base) {
   pir::IrContext* ctx = pir::IrContext::Instance();
